server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["jp calvo <calvojp92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
email-validator = "^1.1.3"
python-multipart = "^0.0.5"
python-dotenv = "^0.19.2"
redis = "^4.0.2"
httpx = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.12.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
